# api/openapi.yaml
openapi: "3.0.3"
# TODO: explore implementation of 3.1 spec w/ overlays, or even different tools.
#       Consider whether using the older OpenAPI spec matters or not.

info:
  title: Typist API
  version: "1.0.0"
  description: |
    Backend for the adaptive typing trainer. The client requests the next
    lesson, the user types it, and the client submits per-item observations.
    The server folds those into the user's competency state (via the adaptive
    engine) and records a session summary.

    Written to OpenAPI 3.0.3 for clean `oapi-codegen` support. Server
    interfaces are generated from this file; handlers implement them.

    Tracking 3.1 implementation as per:
    - https://github.com/oapi-codegen/oapi-codegen/pull/2336
    - https://github.com/oapi-codegen/oapi-codegen/issues/373

    Note: the internal sshd-only token-exchange endpoint (ADR 0007) is
    deliberately omitted. It is not part of the public contract.

servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://typist.gyarmathy.co
    description: Homelab deployment

security:
  - bearerAuth: [] # default; public ops override with: 'security: []'

# Phase 4 work: define paths as the handlers are built.
# Potential initial surface:
#
#   POST   /auth/register
#   POST   /auth/login
#   POST   /auth/refresh
#
#   GET    /progress              - current competency snapshot
#   GET    /lessons/next          - generate next lesson
#   POST   /sessions              - submit a completed session result
#   GET    /sessions              - list past sessions (paginated)
#
# Generate server interfaces with oapi-codegen and implement them in
# each domain package's handler.go.

# NOTE: on error media types: 404 (unknown path) and 405 (wrong method) are
# emitted by Go's http.ServeMux before any handler runs, so they carry the
# stdlib's text/plain body, NOT application/problem+json. They are therefore
# not modelled as problem+json responses below. See ADR 0019.
paths:
  /healthz:
    get:
      summary: Liveness check
      description: Always 200 while the process is serving; does not touch the database.
      operationId: getHealthz
      security: [] # public endpoint
      responses:
        "200":
          description: Service is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /readyz:
    get:
      summary: Readiness check
      description: >
        Returns 200 when the database is reachable within 2s, otherwise 503.
        Used by orchestration to gate traffic until dependencies are healthy.
      operationId: getReadyz
      security: [] # public endpoint
      responses:
        "200":
          description: Service is ready (database reachable)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "503":
          description: Service not ready (database unreachable)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /auth/register:
    post:
      summary: Register user
      description: >
        Returns 200 when the user is successfully registered in the database,
        400 on malformed request (invalid arguments), 409 on conflict (user
        already registered).
      operationId: registerUser
      security: [] # public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: User successfully registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Registration request malformed / invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: User identifier already registered.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /auth/login:
    post:
      summary: Login user
      description: >
        Returns 200 when the user is successfully logged into the system,
        400 on malformed request (invalid arguments), 401 on invalid
        credentials.
      operationId: loginUser
      security: [] # public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: User successfully logged in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Login request malformed / invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Invalid credentials.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /progress:
    get:
      summary: Return user's current competency state.
      description: >
        Returns the user's competency state. Body shape is provisional; finalised
        in phase 4 when the adaptive engine defines the canonical schema.
      operationId: getProgress
      responses:
        "200":
          description: Placeholder for JSON competency state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competency"
        "401":
          description: Unauthorised, not authenticated.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

components:
  securitySchemes:
    bearerAuth:
      description: Bearer token using a JWT.
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      description: Request to register user.
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: Email string.
          format: email
        password:
          type: string
          description: Plaintext password string.
          minLength: 8
          maxLength: 100
    LoginRequest:
      type: object
      description: Login request.
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: Email string.
          format: email
        password:
          type: string
          description: Plaintext password string.
          minLength: 8
          maxLength: 100
    TokenResponse:
      type: object
      description: JWT response, issued on successful register or login.
      required:
        - token
        - token_type
        - expires_in
      properties:
        token:
          type: string
          description: JWT string.
        token_type:
          type: string
          description: Always 'Bearer'.
        expires_in:
          type: integer
          description: Seconds until token expiry.
    Competency:
      type: object
      description: Placeholder for JSON competency object.
    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    Problem:
      type: object
      description: >
        RFC 7807 problem detail, served as application/problem+json. Mirrors
        internal/platform/httpx.Problem. Additional extension members may be
        added later without breaking clients.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          default: "about:blank"
          description: URI identifying the problem type; "about:blank" means "see status".
        title:
          type: string
          description: Short, human-readable summary, stable for a given type.
        status:
          type: integer
          format: int32
          description: HTTP status code, duplicated in the body per the RFC.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: Identifies this occurrence; set to the request ID.